Allow to force restoring the neck in `MutationsHelper.restoreNeck()`
More as a failsafe it is now possible to enforce restoring the neck (e. g. just before a neck TF to a different type). If its being enforced it won't count as a change. In addition to that, the check for changes vs. changeLimit has been removed from that method. If the neck changes from one type to another it should count as a single change.