Move WING_TYPEs from appearanceDefs.as to BodyParts.Wings
### POC: Move appearanceDefs to BodyParts
This is just a POC, how we could slowly but sure make appearanceDefs.as deprecated and maybe remove it from the codebase completely at some point.
Since this POC even affects 64 files, we should IMHO do this in a slower pace.
e. G.: Leave appearanceDefs.as untouched at first. Don't do the find and replace in all files directly, but step by step.
Again: This is a POC. Hance why I've converted the WING_TYPE-consts completely.