Updated the dragon tail and implemented (dragon) tail slam attack
### Gameplay changes
The dragon tail now has a different and more fitting TF blurb and appearance description. In addition to that, the dragon tail now has an unique 'Tail Slam' attack, dealing quite some damage with a chance to stun the opponent and/or let it bleed for three rounds.

### Internal changes
- Added the following set of parser tags:
  - [monster.short]
  - [monster.a]
  - [monster.capitala]
  - [monster.pronoun1]
  - [monster.pronoun1caps]
  - [monster.pronoun2]
  - [monster.pronoun2caps]
  - [monster.pronoun3]
  - [monster.pronoun3caps]
- Fixed a small bug, where the bleed damage on the player displayed the wrong value.

### Notes
If you think, the new attack is unbalanced: The mechanics can always be changed later on.

### Credits
Attack texts, TF and appearance blurbs written by Ein.