Make transformation RNG less nasty
I've created the function `MutationsHelper.tfChance(min:int, max:int):Boolean` as a wrapper around `rand(x) === 0` for TF classes and functions.
It has a min and max value to make it more likely to have at least one mutation, when consuming transformatives.
I've used Reptilum as an example, but this could/should always be ported to other transformatives to make the RNG less nasty.
changes in Reptilum are:
- Replaced `rand(3) === 0` with `tfChance(2, 3)`
- Replaced `rand(4) === 0` with `tfChance(2, 4)`
- Replaced `rand(5) === 0` with `tfChance(3, 5)`
In other words: its always min = the old chance - 2 (the minimum value being 2) and max = the old chance.

Note, that tfChance is not failsafe, so please make sure, that min is always smaller, than max. And with 'always smaller' I mean: Don't replace for example `rand(2) === 0` with `tfChance(2, 2)`, because that would be quite stupid ...