Simplified special eyes calculations
No need for a separate 'has four spider eyes'-check, since two spider eyes are not intended and should never happen through normal gameplay and would be fixed upon loading a savegame.

In addition to that I've cleaned up the eyeCount-section in `Saves.as` upon loading a save file and the eyeCount fix for spider eyes I've mentioned above.